Refactor actions to standalone `Action` component and update references
