+++ resources/js/types/Action.ts
@@ -1,0 +1,13 @@
+import { buttonVariants } from '@/components/ui/button';
+import type { VariantProps } from 'class-variance-authority';
+import type { icons } from 'lucide-react';
+
+export type Action = {
+    className?: string;
+    label: string;
+    href: string;
+    target?: string;
+    variant?: VariantProps<typeof buttonVariants>['variant'];
+    icon?: keyof typeof icons;
+    as?: 'button' | 'a';
+};
